requirements.txt
absl-py==1.4.0
accelerate==0.25.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
ansi2html==1.9.1
antlr4-python3-runtime==4.9.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
av==11.0.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
bitsandbytes==0.42.0
blessed==1.20.0
BLEURT @ git+https://github.com/google-research/bleurt.git@cebe7e6f996b40910cfaa520a63db47807e3bf5c
bleurt-pytorch @ git+https://github.com/lucadiliello/bleurt-pytorch.git@279ca1bb4106bde5a89f0f82723197e23d8446cb
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
cmake==3.27.1
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
confection==0.1.4
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.8
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.20.0
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl#sha256=d88c737eb7eb766f730f6a2dcb99dfcdb81623e1e0d89a9c638a2182ac19c52e
debugpy @ file:///croot/debugpy_1690905042057/work
decorator==4.4.2
decord==0.6.0
diffusers==0.14.0
dill==0.3.8
easydict==1.11
einops==0.6.1
einops-exts==0.0.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
evaluate==0.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fairscale==0.4.13
faiss-gpu==1.7.2
fastdtw==0.3.4
fastjsonschema==2.19.1
filelock==3.12.2
Flask==3.0.0
flatbuffers==24.3.25
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.1.0a0
gpustat==1.1.1
grpcio==1.56.2
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.5
hyperframe==5.2.0
idna==2.10
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1691424382338/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1683289033986/work
ipywidgets==8.1.1
itsdangerous==2.1.2
jax==0.4.13
jaxlib==0.4.13
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1690896916983/work
Jinja2==3.1.2
jiwer==3.0.4
joblib==1.3.2
JPype1==1.4.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1669775081186/work
jupyterlab-widgets==3.0.9
keras==2.13.1
kiwisolver==1.4.4
konlpy==0.6.0
langcodes==3.3.0
libclang==18.1.1
lightning-utilities==0.9.0
lit==16.0.6
lmdb==1.4.1
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mediapipe==0.10.11
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
murmurhash==1.0.10
nbformat==5.9.2
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
nltk==3.8.1
npy-append-array==0.9.16
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
peft==0.7.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
portalocker==2.7.0
preshed==3.0.9
proglog==0.1.10
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.25.4
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-lightning==1.8.6
pytube==15.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
rapidfuzz==3.9.6
referencing==0.32.1
regex==2023.6.3
requests==2.32.3
requests-oauthlib==1.3.1
responses==0.18.0
retrying==1.3.4
rfc3986==1.5.0
rouge-score==0.1.2
rpds-py==0.16.2
rsa==4.9
sacrebleu==2.2.0
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.4.0
sniffio==1.3.1
sounddevice==0.4.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf-slim==1.1.0
thinc==8.2.2
threadpoolctl==3.2.0
timm==0.4.12
tokenizers==0.19.1
torch==2.1.2
torch-tb-profiler==0.4.1
torchaudio==2.1.2
torchmetrics==1.0.2
torchvision==0.16.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.66.5
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers @ git+https://github.com/huggingface/transformers.git@691586b0dcd503afb2a7fc878a408f086e46288b
triton==2.1.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
weasel==0.3.4
webdataset==0.2.86
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.16.2
